Refactor : 파일 구조개선
